collectors/python-snmp-collector/requirements.txt
pysnmp==5.1.0
pyasn1<0.5.0
pysmi<1.3.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.1
structlog==23.2.0
